[libs] E: Moving sched to syscall
This commit moves the sched module from the posix crate to the syscall
crate.
