+++ src/libs/syscall/src/sched/bindings/mod.rs
@@ -5,14 +5,4 @@
 // Modules
 //==================================================================================================
 
-cfg_if::cfg_if! {
-    if #[cfg(feature = "syscall")] {
-        mod syscall;
-        pub use self::syscall::{
-            sched_yield,
-        };
-    }
-}
-
-#[cfg(all(feature = "syscall", feature = "staticlib"))]
-pub mod bindings;
+pub mod sched_yield;
